Add mutex and trace profiles to serviceability, handle unique
The current behavior of writing the file to the current directory
is convenient but doesn't allow parallel profiles to be gathered
for things like `openshift-tests` which uses child processes, or
short lived commands that you want to profile in a loop.

Callers may now append `-tmp` to the mode and profiling will write
to the temp directory which will create a `profile-RANDOM/mode.pprof`
file that is unique across runs.