pkg/assets/create: Allow mid-create cancels
Passing the context into create() allows callers to cancel without
waiting for all of the manifests to be attempted.  At the default 5
requests per second, the installer's 30+ manifests would take 6+
seconds in the "negligable network time" best-case, and we can be more
responsive to cancels than that.

There's a bit more of a dance around lastCreateError now that create
can also return boring canceled and timed-out errors, while we still
want to prefer the last actual manifest-creation failure.