Fix API registration for 1.16
`APIService.Spec.Service.Port` was added in
1.15 (https://github.com/kubernetes/kubernetes/pull/74855) and
defaults to 443. This was causing api registration to continually
update APIService resources due to the expected form lacking a port
and the serialized form having a default value of 443, resulting in
thrashing and API instability. The fix is to not consider the service
port when determining if an APIService is current.