pkg/operator/status: Add configurable inertia
Instead of hard-coding a two-minute inertia, allow the
controller-creator to set a custom inertia function.  The previous
value was chosen to work fairly well for most cases, but there was
contention between important conditions where 2m was pushing too long
and less severe conditions where 2m was short enough to catch slow
launch-time initialization that triggered distracting Degraded flaps
(e.g. the 2m10s missing ConfigMaps in [1], where the bulk of that 2m
was taken by other kube-apiserver initialization actions).  With this
commit, the kube-apiserver can say "3m of missing ConfigMaps is ok,
but only accept 1m of SomeOtherThing before going Degraded=True" or
whatever.  NewClusterOperatorStatusController now sets a degraded
inertia function to match the old default, but callers who want a
different degraded inertia can call WithDegradedInertia to alter their
controller's degraded inertia function.

The Inertia function takes the full condition instead of just the type
in case folks want to switch on some other property, such as the
Reason string.  Passing by value, instead of using a pointer, seems a
bit inefficient.  But it does protect you from malicious Inertia
implementations that want to mutate the passed condition (doesn't seem
very likely), and it also matches the existing internalUnionCondition
implementation which has a lot of by-value copies in its internal
[]OperatorCondition slices.

The distinction between unionCondition and unionInertialCondition was
small, so drop it here and pass all conditions (not just Degraded
types) through, setting nil inertia for non-Degraded types.  It's
possible that these will grow their own WithWhateverInertia helpers in
the future if we find a use-case for them.  If/when that happens, we
might also grow WithInertia that sets the same Inertia function for
all per-type properties.  But WithDegradedInertia is suitably
namespaced so callers who use it now will not break if we decide to
add inertia to other condition types in the future.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1775922