Set a random serial number for signing certificate templates
Previously a serial number of 1 was used when generating signing
certificates. If such a certificate was ever rotated, the new signing
certificate would have the same issuer+serial number which violates
the rfc. Setting a random serial number ensures that signing
certificates can be uniquely identified for a given issuer when rotated.