Update ApplyDaemonSet to rely on a hash of the spec
This removes the need for the caller to be able to force a
rollout. Any state that is not present in the spec should instead be
added as an annotation so that a rollout will occur when the external
state changes.

This is an intentional breaking change to encourage switching to the
revised function. Callers who want to continue using the
previous (deprecated) implementation will need to change the name of
the function they are calling to ApplyDaemonSetWithForce.