Fix staticpod controller to use service ca from prometheus
Previously the staticpod controller was configuring the service
monitor to source the service ca from the service account token
configmap. As of 4.5 the service ca is no longer set in token
configmap, so its necessary to source the service ca from the
prometheus configmap as per the example of most operators.