Fix crypto configuration for golang 1.14
The 1.19.0 bump tried to ensure library-go compatibility with the
addition of a _SHA256 suffix to the following cipher suites:

 - TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256
 - TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256

The naive approach taken by the bump's test fixes was to include both
the new and old names in the cipher map, which satisfied the check in
TestConstantMaps that all golang cipher suites were mapped to a name.

However, CipherSuitesToNamesOrDie was not tested, and when called by
kube-apiserver would panic due to the suite ids of the ciphers in
question mapping to both the old and new cipher names.

This commit ensures CipherSuitesToNamesOrDie is tested to not panic
when called with the default set of ciphers, and ensures that
attempting to conver the suite ids in question always returns the new
name.