controller factory: add informers with filtered events
Allow filtering the events that come into informers so that
only the interesting objects trigger a sync call. This is
useful for controllers that take a certain resource type but
only want to react to the resource changes when a condition
is met, e.g. an annotation is set to a certain value.