add a check for an unhandled error in ApplyServiceMonitor
If ApplyServiceMonitor fails to Get() an existing resource because it
is not there, it correctly handles the error and creates the
resource. However, if Get() fails for some other reason the error was
left unhandled, causing a panic later when a pointer is
uninitialized. This change handles the error to avoid the panic.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>