improve DefaultClientDialContext
wrapping the DialContext method and setting the socket options on a copy of the file descriptor is not deterministic.
as per documentation attempting to change properties of the original using this duplicate
may or may not have the desired effect.

instead we should set the options on the original file descriptor.