Add IsCloudProviderExternal helper for operators
This will allow operators to determine whether or not they should be
using internal or external cloud providers.

We want to be able to have this logic centrally so that we can restrict
the feature gate usage only to platforms that have external provider
support, without duplicating this logic in multiple locations.