Allow missing instance in OperatorClient
Certain operators (esp. those managed by OLM) should work well without any
CR created for the operator. It should do nothing, just sit down and wait
for the CR to be created.

Therefore allow opt-in for "fake instance" mode. In this mode, return a fake
operator Spec + Status with ManagementState: Unmanaged, so the rest of library-go
does nothing and does not log NotFound in weird places.