Force-set ManagementState: Removed when operands should be removed
If an operator is removable and it uses finalizers to remove its operands,
allow ManagementStateController to set the ManagementState from Managed to
Removed when it's time to start removing the operand.

This will simplify other controllers, they need to react to
ManagementState: Removed without checking finalizers and/or
DeletionTimestamp.

Other operators that don't use Removed state or don't use Finalizers
should not be affected by this change.