provides a precondition checker determines if encryption controllers should synchronise
Synchronizing encryption controllers is expensive because they pull data directly from the servers to get the most recent data.
This PR provides a checker that uses the cache for gathering enough data to determine if there is some work to be done.
This helps to avoid sending requests to the API servers if there is no work to do.