extends the maximum time the tests will wait for KAS to reach the same revision
Previously the maximum time was 15 minutes, now it is 22 minutes.

I'm increasing the timeout to fix failing tests on AWS platform.
There is a known issue on that platform with the NLB, it needs way more time to remove an unhealthy instance from the service.

That time has increased from 70 seconds to 210 seconds.
That means the time needed to wait for all instances has increased at least by 7 minutes.