extends the threshold for considering the servers stable
Previously the Kube APIs had to stay on the same revision for 3 minutes to be considered stable.
This is no longer true on AWS platform. The maximum time a server needs to shut down is 270 seconds + the time needed to start the new server.
Thus I am m increasing the threshold to 6 minutes to fix failing tests on AWS platform.