imageutil: New helper for oc adm must-gather / oc debug
The ResolveLatestTaggedImage utility method was intended for callers
that needed to wait for an image import to succeed (trigger controllers
and build controllers that need the image SHA). A number of use cases
exist where import is not required, such as oc debug and must-gather
where the qualified tag is sufficient. Since the utility method might
confuse a caller, create a new utility method intended for the three
most common use cases:

1. You need to get the user's *current* desired intent (oc adm release
   new) and want to error out if the import has errors or isn't
   complete
2. You need to get a pull spec but you don't care if it's out of date
   (oc debug / oc adm must-gather)
3. You need to get a pull spec and you must have imported the image

Add a new ResolvePullSpecForTag method that debug/must-gather will
use.