CSI: Unify assetFuncs for all controllers
Now there is some inconsistency between CSI controllers,
some of which require the asset function to return an error
and others to panic immediately.

Due to the transition to embed module that provides only a
function that returns an error[1], we need to unify our interfaces
for the controllers.

This commit does this by requiring that all controllers now accept
a function that returns an error. For controllers that should have
panicked, we do an explicit check in the code.

[1] https://golang.org/pkg/embed/#FS.ReadFile