encryption precondition checker must indicate work to do when an operator is degraded
The precondition checker will run the encryption controllers will be run when the caches haven't been synchronized.
During synchronization, the controllers send live requests to the API server.
A single failed request can put a controller into a degraded state.
Thus the precondition must take into account the status of the operator and let the controller run when it is in a degraded state.