[staticresourcecontroller]: Conditional decorators
This commit adds conditionals on which resources can be
created/deleted. As of now, static resource controller
is pretty static in creating resources. This commit
ensures that the static resources are created/deleted
based on certain conditional functions. This brings in
needed dependency on other resources and ensures
static resource controller can react to it.