introduces the startup monitor
The main purpose of the new app is to provide a fallback mechanism that will make the operand to revert to the last healthy revision when the new revision fails.
The fallback mechanism is opt-in by the operator, and kube-apiserver and etcd operators (and potentially kcm and ks) have to consult the deployment topology in the infrastructure resource to decide.

For more details see: https://github.com/openshift/enhancements/pull/833

Co-authored-by: Dr. Stefan Schimanski <stefan.schimanski@gmail.com>
Co-authored-by: Lukasz Szaszkiewicz <lukasz.szaszkiewicz@gmail.com>