changes the installer-pod to write optional pods
up until now, the installer pod expected only a single pod manifest in the configmap (the cm is specified by the --pod flag)
since we are planing to add a startup monitor for assessing operand's health check we are extending the installer
to take into account optional keys that might contain other pods.

the new pods will be written to the same locations as the main pod (revisioned, manifests).
the name will be derived from the name filed and suffixed with "-pod.yaml"