startup-monitor: allows for logging to a log file in addition to stderr
since kubelet doesn't respect the restart policy for static pods
we need to store the logs in a file for future troubleshooting

this PR introduces a new flag that allows specifying a log file for the logs.
the file will be rotated on a monthly basis and we will retain only the last 3 files.