Try another registry if blob is not found
ServeBlob uses only the first connected registry, but it does not check
that the registry has needed data. If the blob is absent of this
registry or the registry rejects the request, the client can safely make
another attempt with another registry.

Open also has a problem: it just creates an in-memory object, but the
connection to the registry is established only on the first Read.