installerpod: do not copy resources if a target directory with a higher revision already exists
It is possible to get to a situation when an installer pod of a given
revision (e.g. 6) is requested to be deleted and new installer pod
of a higher revision (e.g. 7) is created. Yet, the container engine
does not delete the installer container of revision 6 before
installer container of revision 7 gets to run and completes. In which
case resources of revision 7 are replaced by resources of revision 6
even though the installer of revision 7 completes successfuly.
The installer controller does not expects this situation to happen
and indefinitely waits until the revision 7 static pod appears.
However, this condition never becomes true.

Also, always acquire a lock to make check for the revision and copying
the resources an atomic operation.