encryption revisionedpod: tolerate pods in Succeeded or Failed phase only if their revision <= the running pods
It turns out that when a node dies or is disconnected from the rest of the cluster, Kubernetes (kubelet) applies a policy for setting the phase of all Pods on the lost node to Failed.
Pods in the Failed state cannot transition back to Running state.
In addition, failed Pods are note GC'ed. The API objects remain in the cluster's API until a human or controller process explicitly removes them.
Thus we need to tolerate pods in Succeeded or Failed phase only if their revision <= the running pods. Otherwise we won't make any progress.

See the following for more:
 https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/
 https://bugzilla.redhat.com/show_bug.cgi?id=2000276