staticpod builder: initialize operand label selector separately
Not every controller consumes operand pod label selector when the StartMonitor is enabled.
Also, the operand pod label selector is now required by the guard
controller since not every operator has the static pod name
equal to the app label. E.g.

| operand | label | static pod name |
| ------- | ----- | --------------- |
| kube-scheduler | app=openshift-kube-scheduler | openshift-kube-scheduler |
| kube-controller-manager | app=kube-controller-manager | kube-controller-manager |
| kube-apiserver | app=openshift-kube-apiserver | kube-apiserver |

In case of KS and KCM both app label and the static pod name are equal.
Whereas in the case of KA, the app label differs from the static pod
name. Thus, it is more practical to provide the operand pod label selector rather than rely on a specific label.