pkg/verify: Expose underlying signature errors
Sometimes we fail to verify because we find signatures we don't trust.
Sometimes we fail to find signatures at all, either because the
backing stores have no signatures, or because we fail to contact the
backing stores.  With this commit, we pass all that useful debugging
context on to the caller.  Not in the string, because we don't want to
overwhelm all callers, but they can get it via Unwrap [1] so they can
show the details where they have space to do so.

[1]: https://pkg.go.dev/errors#Unwrap