Fix ApplyService to detect new Service spec
When an operator needs to update service.spec, ApplyService should
apply it. We detect change of the spec by using an annotation with the old
service.spec required by the operator and comparing it with the current
operator's request.

Keep the existing heuristics of user changes (spec.type, spec.selector) in
place.

This code expects that the API server restores any immutable service.spec
fields, like clusterIP.