pkg/verify: Include Signatures error in aggregated error
We can hit this case if lookup fails (e.g. because it takes too long
to retrieve over HTTP, and the Context times out).  Instead of failing
with just "context deadline exceeded", timestamp this error, roll it
in with the others, and fall through to consider any remaining
verifiers who have not yet been satisfied.