pkg/verify/store/parallel: Best-effort, non-blocking response write on context cancel
From the spec [1]:

  If one or more of the communications can proceed, a single one that
  can proceed is chosen via a uniform pseudo-random
  selection. Otherwise, if there is a default case, that case is
  chosen. If there is no default case, the "select" statement blocks
  until at least one of the communications can proceed.

So in this callback, we have been called by Signatures, and have a
signature/errIn we'd like to pass back through responses.  But maybe
responses is full.  In that case, we try to block on the write, in
case we get some space to write later on.  But we don't want to block
forever.  Previously, the parallel ctx.Done() would bail us out of a
too-long wait (good), but in cases where it won the select
random-choice we would discard the response data we'd been trying to
send (bad).  With this commit, the ctx.Done() case gets a single,
non-blocking write attempt, so we can pass along the signature/errIn
information if there happens to be space.  If not, there will be
plenty of earlier responses in that channel that we'll be able to
process later.

While I'm touching things here, pivot from 'true, nil' to 'false,
ctx.Err()' to show that we're a bit grumpy about being canceled.  The
errorChannel collector is going to silently discard the Canceled error
when it's aggregating errorChannel content, but the presence of the
error will make it less likely that later logic pivots misinterpret
this as "successful validation".

[1]: https://go.dev/ref/spec#Select_statements