add ability to detect vsphere csi migration
Due to changes in upstream kubernetes, the vSphere in-tree cloud
provider will need to be used in situations where the in-tree storage is
still in use. Because OpenShift 4.13 is enabling CSI migration by
default for vSphere, there will need to be a regression method for the
CCM as well. This change transforms the IsCloudProviderExternal function
to detect this condition.

references:
https://groups.google.com/g/kubernetes-sig-storage/c/wvJpNDBS9vU
https://github.com/kubernetes/kubernetes/pull/116342
https://issues.redhat.com/browse/STOR-1265