library-go: GetName must not return value starting from the dash
If the base is long enough and the suffix is of 243 symbols, `GetName()` calculates `baseLength` equal to zero and proceeds with construction of result value:
```
	prefix := base[0:baseLength]
	// Calculate hash on initial base-suffix string
	return fmt.Sprintf("%s-%s-%s", prefix, hash(base), suffix)
```
Here the `prefix` is the empty string and the return value statrts from the dash (`-`). It violates the standard: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/

The problem is easy to fix: we must fall back to truncated_base+hash if `baseLength` is zero.