OCPBUGS-15390: Support Group/Version format in manifest override
Prior to openshift/cluster-version-operator#689 the 'group' field in an
override was ignored. During that time, the documented way to use this
field was to specify a Group/Version (e.g. "apps/v1") matching the
apiVersion field in the manifest.

After that change, only an exact match of the override's 'group' field
and the Group part of the GVK is allowed, so the 'group' must be set to
e.g. "apps" and never "apps/v1".

This change allows both "apps" and "apps/v1" formats to work.