USHIFT-1593: Regenerate certs on hostname changes
When generating certificates there is a check to see if the cert
already exists. If it does, it compares the existing IPs and DNSs
in the cert against the hostnames it should have. If there is one
missing entry from the hostnames in the cert it is regenerated. If
not, then the cert is left untouched.
There is no check to see if the entries in the cert are present in
the hostname list, leaving room for outdated IP/DNS entries to remain.