Use k8s.io/utils/ptr instead of deprecated pointer
Also remove unused pointer manipulation functions.
resourcemerge.BoolPtr is used elsewhere so it's only deprecated for
now.

In the common "modified" boolean pattern with EnsureObjectMeta, it's
simpler to declare a boolean and pass its pointer to EnsureObjectMeta,
rather than declaring a pointer to a boolean with a pointer wrapper
function; this approach is used in many cases instead of a blind
replacement of resourcemerge.BoolPtr with ptr.To.

Signed-off-by: Stephen Kitt <skitt@redhat.com>