NO-JIRA: add exponential backoff to static pod installer
This overrides the default rate limiting for the controller to be less frequent.
Additionally, the base controller now rate limits informer invocations
in the same way.

We observed the installer controller to hot loop, especially when revisioned
resources are still missing to be copied while a revision is rolled out, which
causes CI failures with pathological events.

Signed-off-by: Thomas Jungblut <tjungblu@redhat.com>