pkg/verify/configmap: Aggregate errors in loadArmoredOrUnarmoredGPGKeyRing
Previously, when both ReadArmoredKeyRing and ReadKeyRing failed, we'd
only return the latter error.  But in [1], ReadArmoredKeyRing errors
like:

  openpgp: invalid data: user ID self-signature invalid: openpgp: invalid signature: RSA verification failure

pointed at the SHA-1 self-signature issue, but were masked by
ReadKeyRing's:

  openpgp: invalid data: tag byte does not have MSB set

and required additional work on out-of-cluster reproducers.  With this
commit, we'll return an error aggregating both underlying errors on
failure, to make future debugging easier.

[1]: https://issues.redhat.com/browse/OCPBUGS-35528