avoid setting .status.ObservedGeneration in tests
We are currently setting this field in the initial objects and only
verifying its presence after the controller sync. However,
controllers are not mutating this field, as they shouldn’t, since it
could cause race conditions. This update removes the unnecessary
setting of the field in tests, ensuring the tests better reflect
expected controller behavior.