Fix "live get after write" in static pod installer controller.
The static pod installer controller builds an apply configuration for the status of the nodes it
manages based on potentially-stale state from an informer cache. The controller is written to assume
that it has observed the effect of its own previous writes. If the cache is stale, this assumption
can be violated which results in unpredictable installation decisions.

A mechanism was recently introduced requiring the installer controller to wait for its lister to
catch up to the latest version after performing a write. This mechanism did not work because it
depended on keeping state across calls to the Sync method, and because Sync has a value receiver,
field writes were not visible on subsequent calls.