certrotation: don't update metadata in RefreshOnlyWhenExpired mode
There are two copies of this controller running in the cluster:
one in the operator with RefreshOnlyWhenExpired=false and one leader in
kube-apiserver sidecar with RefreshOnlyWhenExpired=true. This method
allows us to generate certs in operator before kube-apiserver pods are
created - and regenerate expired certs after the cluster was shutdown.

However, regardless of the mode, this controller applies metadata changes.

So if an update introduces a change in some annotations these controller
copies are racing, as there are different versions of the controller
running - an updated one in the operator and outdated in kube-apiserver
sidecar.

This change ensures that controllers in
RefreshOnlyWhenExpired=true mode don't update metadata. The primary goal
is to have certificate contents fresh enough to run the control plane -
and the operator will take care of the secondary things like
secret/configmap metadata.