Only check deployment time when remaining in Progressing state
Once a deployment takes more than 15 minutes to come up, the owner
operator can never get out of the progressing state.

This check was introduced in #1914

However, it is positioned in a way that will never allow a cluster
operator to get out of the progressing state, it will always return
an error if the deployment has taken too long.

Move the check to occur only when the operator remains in the
progressing state. If the operator is exiting the progressing
state, then don't do the check.

Signed-off-by: Todd Short <todd.short@me.com>