certrotation: exit when received cabundle is empty
When different controllers update the same CA bundle a conflict may occur. In that case we return an
empty set of certificates and empty error.

The sync procedure should exit if either non-empty error received or cabundle is empty, as this ca
bundle is later used to detemine whether target cert can be verified. If conflict occurred during CA
bundle update we should restart the sync worker again - otherwise this may lead to target
certificate unexpectedly regenerated. When service network target certificate is being updated it
may be fatal, as the operator will lose its connection to API server and won't be able to receive an
updated certificate.

Co-Authored-By: Masaki Hatada <mas-hatada@nec.com>