Fix CSI node DaemonSet progressing
The Progression condition should be set only when the DaemonSet has a new
content and is actively rolling out its update. After all replicas are
updated, the Progressing condition should be false, even when some pods are
missing. E.g. because a node is drained, something evicted them or so on.

To find out if a DaemonSet has finished roll out (= stable), store the generation
of last stable DaemonSet in a DaemonSet annotation.

If the DaemonSet misses a pod, check its annotation. Is the current
DaemonSet generation equal to the last stable generation? Then it's not a
reconfiguration of the CSI driver and do not set Progressing condition.
Is the current DaemonSet generation different? The operator has changed the
DaemonSet spec, i.e. reconfiguration is in progress, i.e. set the
Progressing condition.

There is still a case when user changes the DaemonSet spec and thus bumps
the generation. Progressing condition will be set. We consider it
acceptable.